+++ packages/utils/icons/ArrowRight.tsx
@@ -1,6 +1,9 @@
 import { IconPropsType } from "./type";
 
-export const Copy = ({ size = 24, color = "currentColor" }: IconPropsType) => (
+export const ArrowRight = ({
+  size = 24,
+  color = "currentColor",
+}: IconPropsType) => (
   <svg
     width={size}
     height={size}
@@ -11,7 +14,6 @@ export const Copy = ({ size = 24, color = "currentColor" }: IconPropsType) => (
     strokeLinecap="round"
     strokeLinejoin="round"
   >
-    <rect width="14" height="14" x="8" y="8" rx="2" ry="2" />
-    <path d="M4 16c-1.1 0-2-.9-2-2V4c0-1.1.9-2 2-2h10c1.1 0 2 .9 2 2" />
+    <path d="m9 18 6-6-6-6" />
   </svg>
 );