+++ packages/utils/icons/Success.tsx
@@ -1,6 +1,9 @@
-import { IconProps } from "./type";
+import { IconPropsType } from "./type";
 
-export const Success = ({ size = 24, color = "currentColor" }: IconProps) => (
+export const Success = ({
+  size = 24,
+  color = "currentColor",
+}: IconPropsType) => (
   <svg
     width={size}
     height={size}