+++ packages/utils/icons/index.ts
@@ -1,13 +1,31 @@
+import { ArrowRight } from "./ArrowRight";
+import { Bun } from "./Bun";
 import { Copy } from "./Copy";
 import { Error } from "./Error";
+import { GalleryVertical } from "./GalleryVertical";
+import { GalleryVerticalEnd } from "./GalleryVerticalEnd";
+import { Github } from "./Github";
 import { Info } from "./Info";
+import { Npm } from "./Npm";
+import { Pnpm } from "./Pnpm";
+import { Storybook } from "./Storybook";
 import { Success } from "./Success";
 import { Warning } from "./Warning";
+import { Yarn } from "./Yarn";
 
 export const Icons = {
+  ArrowRight,
+  Bun,
   Copy,
   Error,
+  GalleryVertical,
+  GalleryVerticalEnd,
+  Github,
   Info,
+  Npm,
+  Pnpm,
+  Storybook,
   Success,
   Warning,
+  Yarn,
 };